Add functional tests for user config file
This commit adds additional testing to cover the newly added user config
file feature from #139. This includes various permutations of CLI and
config file options to test for the correct operational behavior. This
testing managed to uncover a bug with empty config files which was
fixed.

One outstanding issue is the load abbreviated output option which
doesn't fully function. The abbreviated output (ie the '....') doesn't
get displayed, other than that it's fully functional. I assume this is a
bug in the subunit_trace module, so for now those 2 tests are skipped.