Update the package summary in setup.cfg metadata
This commit updates the description of the stestr package. The one in
setup.cfg is quite old back from when we were still deciding on the
scope of the project. We have since updated the descriptions everywhere
else to explain the scope of the project. The setup.cfg summary is what
gets sent to pypi: https://pypi.python.org/pypi/stestr/2.0.0 so we
should update this to make sure it reflects the current scope of the
project more accurately.