Fix a small mistake in manual about test selection
This commit fixes a small oversight in the stestr manual around test
selection. Previously it incorrectly said you use --list/-l to run a
test list, which isn't actually a thing in stestr. (I assume it was a
copy and paste error from ostestr) This really should read use stestr
list, which this commit does.