Fix the manual docs link in README
The stestr README used to include a previous link to the stestr manual
page. This was done using sphinx internal refs, because in a
sphinx doc build this is the cleanest way. However, on github the README
is not rendered using sphinx, it just parses the rst directly and
converts that to html. This means we would have a link that didn't work.
To fix the rendered html output on github this changes that link to just
be an http link to the read the docs page for the stestr manual.