Add usage to manual docs page
This commit leverages the cliff autoprogram documentation sphinx to add
usage sections for each command to the beginning of the stestr manual.
This is a standard thing to list out each option in the cli which we
were missing from the docs. Luckily cliff provides an interface to
generate this so we don't have to write it out manually.