Fix manpage generation
When using sphinx to generate a man page for stestr it was building a
man page out of all the docs in stestr. This worked fine but made a
pretty bad man page. The MANUAL.rst file is meant to be both an HTML
manual for using stestr and also would make a good man page. This commit
adjusts the root doc for manpage generation to be MANUAL.rst instead of
index.rst. This results in a much more sane manpage when running a
command like:

sphinx-build -b man doc/source doc/build/man

At the same time this updates the application description so it no
longer says "stestr - stestr Documentation" and instead uses the
official project description from the github page.