Add option to suppress attachment printing on successful tests
This commit adds a new cli and user_config option to suppress printing
attachments on successful test runs. Originally in subunit_trace stdout
and stderr were unconditionally printed on test success because the
theory was this was expected to be inadvertent and likely a bug so end
users would want to know if their tests were writing to to either
stream. However, this was a faulty assumption in that some test suites
write to either during normal operation and writing the contents on test
success is just polluting the output of a test run. Therefore this adds
an option to let the user configure this functionality to address that
use case.