Add docs and helper script for building man page
Since stestr is a cli tool having a proper man page for it is useful. We
recently spent some time cleaning up our sphinx usage to enable
generating a proper man page by default with the sphinx man builder.
However, this wasn't documented anywhere or clearly explained. This
commit attempts to address this by first adding a small helper script to
generate a man page using sphinx and then also documenting it in the
readme. Hopefully this will encourage packagers to include a man page in
packages for the next stestr release.