Ignore errors on cleanup in test_return_codes
This commit is an attempt to mitigate #162 where the temp directory
removal is failing on windows CI occasionally. This commit sets the
ignore_errors flag on the shutil.rmtree() call which makes any errors
encountered not fatal. While this isn't ideal because it could result
in leaking the temp directories on test runs. But, since we haven't
encountered this failure on any operating system or environment
besides the windows CI in appveyor the risk of that weighed against
the nondeterministic failures blocking changes minimizes that concern.