Uncap pbr
This commit uncaps the pbr version. The PBR windows issue reported
in: https://bugs.launchpad.net/pbr/+bug/1761134 no longer seems to
effect the windows CI jobs. Since this is the only method we have to
test functionality on windows this should be safe. If there are still
configurations that this doesn't work on, they'll have to cap the PBR
version downstream. The requirements.txt still sets a low minimum bound
to enable doing this.