Fix description for the stestr run command
The description for the run command being returned by the
get_description() function was truncated because the string was split
over multiple lines. However, the split was done incorrectly and the
second part wasn't being set. This commit corrects the oversight so the
description is complete in the documentation and --help.