Add error handling for invalid input regexes
This commit adds error handling for invalid input regexes on any of the
test selection mechanisms. Now when an invalid regex is encountered an
error message will be printed and then stestr will exit.

The logic in selection.py needed to be reworked slightly to
differentiate between invalid regex in input filters and the whitelist
file.