WIP: Fix unxsuccess return code
This is commit fixes the response code from runs using subunit trace
with unxsuccess test results. Currently the test result object isn't
properly tracking the unxsuccess tests, so when we check if the run was
successful when it looks for any unxsuccess tests it doesn't see any and
returns 0. Since this is happening in a lower level for the time being
this fixes it at a higher level using subunit_trace to manually count the
unxsuccess tests. This however doesn't fix the case of
--no-subunit-trace, which will still not work because it never calls the
subunit_trace module.