Add support for test class and method by path on no-discover
This commit adds support to the --no-discover/-n flag on stestr run for
specifying a test class and test method when passing in a filepath. This
borrows the syntax from pytest so you use '::' to specify a class and
then method. For example 'path/test_mod.py::TestClassA::test_method_b'
would be used to run only test_method_b in the TestClassA test case in
the path/test_mod.py file.

Fixes #180