Use reported times instead of wall time in subunit-trace
This commit updates the subunit_trace module to leverage the timestamps
reported in the subunit instead of the wall time for the elapsed time
output. When we were using the wall time if we were just reading a
subunit file this would be the time it took to read an process the
subunit data. Which is not the actual run time of the test. By switching
to use the elapsed time between the first start timestamp and the last
stop timestamp we'll always get an accurate value. This does come with
the tradeoff of not counting any setUp before the start time or
tearDown/cleanUp after the last test, since that data isn't in the
subunit stream. But that feels like a worthwhile compromise for having
moderately accurate numbers all the time.

Fixes #119