Fix time measurement for load command too
The stestr load command has it's own code for running subunit-trace
since it deals with the streams at a lower level than other commands.
(it has to handle realtime operations on the streams and multiplexing
them) When the previous commit changed the behavior of the elapsed time
measuring for the subunit_trace module it didn't have an effect on
the load command because of this. This commit fixes the oversight and
mirrors the behavior in the subunit_trace module for the load command.