Introduce parallel-class option
This commit introduces the parallel-class option to the CLI and config
file. The option take priority over group_regex. And if both on the
CLI and in the config are set, we use the option on the CLI not in a
config file.

For example, `--group-regex` on the CLI and `parallel-class` in a
config file are set, `--group-regex` is higer priority than
`parallel-class` in this case.

Fixes #198