Add a better description to README Overview section
This commit adds some text to the README's overview section. Previously
it didn't actually directly say what stestr was, it just starts talking
about being a fork of testr. This cleans this up by describing what
stestr is and how it works before talking about it being a fork of
testr.