Update docs for exit codes with --subunit flags
The --subunit flag changes the exit code behavior in a slightly
counter intuitive manner. Basically instead of exiting non-zero when
there are failure results it only returns non-zero if subunit generation
fails. This is expected behavior, but was not clearly documented which
can cause confusion. It was only mentioned in the description for the
'last' command, and not all the commands where --subunit was available.
This commit attempts to address that by adding clear documentation to
the description string for each command. This means it will be printed
in the summary for '--help' on each command as well as ending up in the
project man page for each command.

Fixes #210