Inline CONTRIBUTING.rst in README.rst and add to docs
We have had a CONTRIBUTING.rst file in the repo for some time (at the
suggestion of Github) but this is not discoverable unless you can find
the repo. As pointed out in issue #218 the repo can be hard to find in
certain cases, so this document isn't always discoverable. In an effort
to fix this, this commit updates the README to include this content.
Additionally it adds the CONTRIBUTING.rst file to the rendered docs.

Fixes #218