Don't install sphinx in python2 jobs
Sphinx >=2.0.0 has dropped support for python2 (and python 3.4) we
already build our docs using python 3 but sphinx was always being
installed in our python2 ci jobs. This commit adds a minimum python
version to the test-requirements.txt file so that we no longer attempt
to install sphinx on an unsupported python version.