Add release automation for pypi artifacts
This commit adds travis jobs on tag releases to build the release
artifacts for pypi. It will generate both an sdist and a universal
wheel and then upload that to pypi. This automates the release process
so that when it's release time we only need to push a tag to github
and everything else is handled automatically.