Mention IRC channel in README
This commit adds a note about the project's IRC channel to the README
file. The channel is very inactive and that's likely because no one
knows that it exists. It's good to have a place to ask a question about
the project in real time, so this commit adds a note about the channel
to advertise that it exists and is a place to discuss all things
stestr.