Fix edge case where there are no test times in output
This commit fixes an edge case with the load command when there are no
test events in the subunit stream being passed to load. If the
pretty_out flag is set for the load command and there is no subunit
output the processing being used to handle the test run timing will fail
because there are no tests. This fixes that by just returning the test
run failed (because there were no sucessful tests run).

Fixes #244