Remove unused bits from test runner
This commit removes some unused bits for old unittest compat and pieces
of the unittest api that do not apply to the internal stestr use case.
Since the internal use case for the test runner used by stestr is a
smaller subset of what the unittest api provides (mostly because stestr
provides that functionality itself) we are able to remove a large chunk
of code from the test runners. Also since some of the code was
originally based on testtools and python-subunit there were lots of
legacy pieces in there for supporting older versions of python which
does not apply anymore, because this runner is only used with python
>=3.5