Use StringIO instead of tempfile
This commit makes unit tests in test_output.py use io.StringIO instead
of tempfile. It can reduce the code (no need seek(0)). And this commit
also fixes inconsistent code in output.py::output_tests() because it
caused an error with StringIO and it should be the same with the later
one.
The last thing is fixing the indentation in output_stream() which was
just weird.