Revert to previous stdout and remove usage modification
Python 3.6 on windows is still failing CI with an error related to an
invalid handle. It's still unclear what could have caused this error in
the pruning that occured earlier. This attempts to adjust the stdout
usage to be as before because that's the only thing changed that relates
to anything with a "handle". At the same time the usage function to add
a custom message isn't needed since the runner won't be manually run we
don't need more details in the test runner usage output.