Add linux azure pipelines
This commit adds a linux (Ubuntu 16.04) testing job strategy to the
azure pipelines. This commit doesn't remove the current travis jobs
because it still works well and we should look at its stability of the
azure one. So, the next step would be removing the duplicated job
definition and implementing coveralls execution in the azure-pipelines.