Remove stray reference to removed short form of --random
In #267 we removed the short form for --random (-r) since this was
duplicated with --repo-type's short form (also -r). Since the 2.0.0
release which introduced cliff usage the -r option only worked for repo
type and would always error when trying to use it like --random. However
that PR neglected to remove a reference to the short form of --random in
the stestr manual. This commit corrects the oversight and removes the
reference.