Add --pdb flag to stestr run
This commit adds a new flag --pdb to stestr running a test id (in a
similar fashion to --no-discover) for when pdb is going to be invoked
from the tests. Pdb has trouble running in situations where another
process is launched (which is pretty much all of stestr, including with
--no-discover). This commit adds a --pdb flag so that you can launch
the test runner without any additional processes being launched. This
will run the tests in a process capture the subunit output to an buffer
and after the tests finish executing will pass that to the load command
so it can write it to the repository and generate subunit-trace or any
other output format (per the set flags). The tradeoff here is that the
output will not be real time from the test instead it will only be
processed until after the test executes.

One thing to note is that if you're using a fixture to capture stdout
that will sometimes take the output from pdb so you will be in the
debugger but get no output. If you plan to debug a test suite with a
fixture make sure to disable that first.

Fixes #14