Add python 3.8 macOS and windows ci jobs
Python 3.8 support was added to stestr in #277, but at that time the
azure pipelines ci environment did not have support for python 3.8 in
the images. So CI for python 3.8 on Windows and macOS was not added at
that time because we use azure pipelines for that. However, python 3.8
has been added to the azure pipelines images now so this commit adds the
missing CI jobs to verify that everything works as expected on python
3.8 for windows and macOS users.

Fixes #280