Treat bytes and strings differently
The 'as_text()' helper function for the 'testtools.content.Content'
class fails if it attempts to decode a binary result. The fact that we
have a binary result is a bug in itself but we need to log what we're
getting to help debug this. Start ignoring these by default but add an
option, '--show-binary-attachments', to allow us to emit these if we
want it for debug purposes.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Related-bug: #1813147