Update test-requirements to be py3 only
This commit updates the test-requirements to remove things no longer
needed in a py3 only world. This involves updating the sphinx version so
that we don't have to explicitly set a python 2 version. It also means
removing the mock library as we can just use mock from stdlib now that
we're only supporting python3.