Change option '--whitelist-file' to '--include-list'
Added the new option while posting a Deprecation Warning for the old one.
The old option will be supported for now but it is soon to be retired...
So for now, both options are equally usable and will have the same effect.

Tests relying on the old option are kept intact, will be removed only when
the option itself is removed. New tests are added to verify that the same
behaviour is afforded if the new option is used also.