Change option '--blacklist-file' to '--exclude-list'
Similarly as done with '--include-list', the old option will work for now
though it is slated to be removed soon - a Deprecation Warning is added to
notify this. The new option '--exclude-list' is now added covering the same
functionality.

Tests are added for the new option in addition to existing tests for the old
option.