Behaviour with a mix of New and Old options
In case the newly added options are used along with the ('soon to be'
deprecated) old options, the value from the new option is given precedence.

Added clarification in favour of this behaviour in the help string of
the old options - this is useful to know, while both the options co-exist!
Also added similar info in the docstring of the Python API.

Added tests to verify this behaviour - these tests need to be in force till
the Deprecation Notice holds!