Remove subprocess usage from discovery
This commit removes the usage of subprocess from the discovery/list test
phase in TestProcessor. Previously to get a list of tests from discovery
we would subprocess out to a new python interpreter calling the subunit
runner with the list command and capture stdout from process and parse
that sububit stream to get the test ids. However, this was not strictly
necessary as we're not doing anything in parallel, so we should be able
to just call the subunit_runner directly inline and use an in memory
buffer instead of stdout. This has the advantage of not interfering
with sys.stdout, in the previous method if anything besides the
subunit_runner was printing to stdout it would cause a parser error
because of the non-subunit data mixed into the stream.