Migrate CI to github actions (#299)
In the recent 3.1.0 release the travis job for uploading to PyPI failed
because of requirements incompatibilities in the travis CI environment.
That coupled with recent changes in billing and free tiers for travis CI
it seems like a good time to change to github actions, which has better
integration with github and better support at this point. It also
enables unifying the multiplatform testing to a single CI system. This
commit moves all the test and release jobs (except for coverage) to
github actions and removes the old jobs.