Report tests stuck as inprogress as failures
In some situations (especially when testing compiled extensions) a test
case can trigger a segfault or some other scenario that crashes python.
When this occurs the worker doesn't send back the second subunit event
that reports the test status and that leaves the test case stuck as
inprogress. This becomes tricky to debug as there is no clear message as
to why the test run failed. While we were correctly not returning a 0
return code in this situation there was no error message about why we
didn't. This commit fixes this oversight by adding a check to the end of
subunit-trace so that we report any tests with an inprogress state and
explain that we're failing because they likely crashed python.