Add support for setting stestr config in a tox.ini file
This commit adds support for setting the stestr config in a tox.ini
file. Using tox to run tests with stestr is a very common pattern and
instead of requiring users create a separate per repo config file we
can just integrate stestr's configuration in the tox.ini config.

Fixes #300