Deprecate the sql repository type
This commit deprecates the sql repository type for future removal in the
next major version release (4.0.0). Included in this deprecation is the
--repo-type argument which only makes sense with >1 repository type. The
sql repository type depends on the subunit2sql project which is not
really active anymore. It never managed to reach feature parity with the
default file repository type and has been listed as experimental it's
whole existence. At this point it's realistically never going to replace
the file repository which was the original long term plan. One note is
that we both emit a DeprecationWarning and print a warning to stderr
when the sql repository type is set. This is because in my local testing
the DeprecationWarning was not actually show to the user, but I left it
in place because it is the standard python mechanism for these things.

Additionally, since we're deprecating the repository type to save some
CI time and complexity I removed the sql repository tests from the
default test run. We still run it in the macOS CI jobs to ensure we
don't regress during the deprecation but we don't need to spend time
running it in every run.

Fixes #307