Split history subcommands into separate Command classes
The previous commit built the 3 history commands as subcommands on the
parser for a single cliff Command class, but this didn't really work
with the cliff model. Since we really only wanted a UI like
`stestr history list` and `stestr history show` we didn't really need to
use subcommands as long as the UI looked like that. Cliff enables this
by having an underscore in the entrypoint for different commands. This
commit splits the subcommands out into 3 separate Command classes which
have entrypoints that result in the desired UI.