Fix typo in README_ja
In #311 the README_ja.rst file was updated to update references to the change
in default branch name, as part of that change the badges were updated to
remove the obsolete badges for the old CI systems we no longer use and update
it to use a badge for github actions. However, there was a copy and paste error
in that link and it was referring to the wrong repository and not rendering correctly.
This commit fixes the oversight.