Rework stestr history list as a cliff lister command
As was pointed out in code review, now that we have separate commands for
history list, history show, and history remove we can leverage more of
the cliff infrastructure for the list command. This commit converts the
list command to be a cliff lister which will default to the same pretty
table output we were previously manually constructing, but also provide
options for additional output types.

Co-authored-by: Masayuki Igawa <masayuki@igawa.io>