Remove deprecated command line flags
In preparation for the stestr 4.0.0 release this commit removes the
deprecated command line options. This includes '--whitelist-file',
'--blacklist-file', '--black-regex' and '--partial'. The selection
options have been deprecated for > 1 year and the '--partial' option has
been deprecated > 4 years and hasn't had any effect in that time. The
4.0.0 release will be primarily to make these deprecation removals
(along with the sql repository type which will be in a separate commit).