Remove deprecated sql repository type and repo_type options
In preparation for the 4.0.0 release this commit removes the deprecated
sql repository type (which is based on subunit2sql) and the repo_type
cli flag and api parameters. Both of these have been deprecated for > 1
year. The sql repository type never reached feature pairity with the
default file repository, primarily because subunit2sql development
stalled (which is mostly my fault for moving on to new projects). As
it was never really a viable user option because of this and
maintaining the dual repository type path was adding additional
development overhead for additional features we were hoping to
implement it was deprecated in preparation for removal in the 4.0.0
release.