Add coverxml task and cover job
This commit adds coverxml task to tox.ini and cover job to main.yml to
enable coverage tracking. We already have it in travis.yml but it seems
it hasn't worked recently. And, it's a good timing to migrate to GitHub
Actions if it works well because we can remove travis.yml just only for
tracking the coverage.