Reuse subunit_trace output logic for load/run
Previously `stestr run` and subunit_trace used slightly different
output and validation logic. run wouldn't catch the case where a test
exited without a status, while code using `last` and `history` would
since they used subunit_trace.

Refactor the subunit_trace output and validation logic and use it for
`load`/`run`.

Converge on the subunit_trace formatted output, in addition to the
summary also show if something goes wrong.

The primary motivation for this change is to better detect when a
subprocess is killed and not all tests are run as expected (segfault,
OOM etc.).