Revert "Fix #330 - Preserve load-list order"
Several issues have been reported about the run order and other odd
behavior in running of tests since the 4.0.0 release including a high
degree of random failures in stestr CI. #331 is the only likely
candidate as to the root cause. In the interest of fixing these issues
for people this commit reverts #331.

This revert was already released as part of 4.0.1. Since the main branch
adopted `black` code formatting recently it was easier to do the
straight revert on the stable branch first as there was conflicts.

This reverts commit ff25c999a59a5d8fa82ed118c53b11ed479de9bd.