Remove check for Python < 3.5
I originally developed this feature when we still supported older
python versions in stestr. The dynamic scheduling feature depends on
functionality added in Python 3.5. Since then the WIP feature branch sat
stale for years since that time we've bumped the minimum version of
Python supported to 3.7 so the runtime check for older python versions
is no longer needed.