Fix config file test mock
This commit fixes the mocking used for the config file tests. It was
previously implicitly defining a runner argument as a mock object which
wasn't a valid choice for a config file to contain after parsing. This
commit just explicitly sets the mock to have the unittest runner set to
avoid failure.